:white_check_mark: refactor rabbitmq test to integration test
